Tried to make decon code handle even larger files.  However:
Oversubscription (i.e. allocating arrays bigger than GPU ram) doesn’t sound possible on Windows. Stinky pants.
http://disq.us/p/1ji735f

Device memory oversubscription is possible for GPUs that have a non-zero value for the device attribute CU_DEVICE_ATTRIBUTE_CONCURRENT_MANAGED_ACCESS.
https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MEM.html

Updated deviceQueryDrv.exe so it prints out the value of CU_DEVICE_ATTRIBUTE_CONCURRENT_MANAGED_ACCESS, which has been NO for all Windows based cards I've checked so far (P6000, Titan V, Titan RTX,...)

+++ RL-Biggs-Andrews.cpp
@@ -65,12 +65,12 @@ void RichardsonLucy_GPU(CImg<> & raw, float background,
 
 
   // "raw" contains the raw image, also used as the initial guess X_0
-  unsigned int nx = raw.width();
-  unsigned int ny = raw.height();
-  unsigned int nz = raw.depth();
+  size_t nx = raw.width();
+  size_t ny = raw.height();
+  size_t nz = raw.depth();
 
-  unsigned int nxy = nx * ny;
-  unsigned int nxy2 = (nx/2 + 1)*ny; // x=N3, y=N2, z=N1 see: http://docs.nvidia.com/cuda/cufft/#multi-dimensional
+  size_t nxy = nx * ny;
+  size_t nxy2 = (nx/2 + 1)*ny; // x=N3, y=N2, z=N1 see: http://docs.nvidia.com/cuda/cufft/#multi-dimensional
 
 #ifndef NDEBUG
 #ifdef _WIN32
